+++ internal/attempts/interface.go
