+++ internal/attempts/model.go
